# add cache for part builder