@timriley @solnic the base branch is not master but `improve-Path#template-performance` 

Here is another PR that applies cache for `PartBuilder#call` method

Is not a huge win, but the performance has improved 😄 

**Before:**
```bash
Warming up --------------------------------------
   action_controller     1.000  i/100ms
            dry-view     1.000  i/100ms
Calculating -------------------------------------
   action_controller      6.420  (±15.6%) i/s -     32.000  in   5.031820s
            dry-view      3.644  (± 0.0%) i/s -     19.000  in   5.215104s

Comparison:
   action_controller:        6.4 i/s
            dry-view:        3.6 i/s - 1.76x  slower
```

**After:**
```bash
Warming up --------------------------------------
   action_controller     1.000  i/100ms
            dry-view     1.000  i/100ms
Calculating -------------------------------------
   action_controller      6.671  (± 0.0%) i/s -     34.000  in   5.097874s
            dry-view      4.236  (± 0.0%) i/s -     22.000  in   5.194745s

Comparison:
   action_controller:        6.7 i/s
            dry-view:        4.2 i/s - 1.58x  slower
```