@timriley 2018-12-21
@GustavoCaso Thanks for taking this on. With this one I was thinking we could probably just cache inside `#resolve_part_class`? Does that yield pretty similar results?