@timriley 2018-12-21
@GustavoCaso Also, I wonder if it might be easier for us to use `dry/core/cache` rather than building our own here