@timriley 2018-12-21
This is superseded by #97, let's move any further discussion there