@GustavoCaso 2018-12-21
@timriley I tried at the beginning with `resolve_part_class` and the results were not as good, but I can change the code run the benchmarks and post the results and we can decide 😄